+++ Changelog.md
@@ -1,6 +1,11 @@
+<a name="0.6.2"></a>
+## 0.6.2 (2019-09-24)
+
+* Added new endpoints for Card processing (invoices, transactions)
+
 <a name="0.6.1"></a>
 ## 0.6.1 (2019-07-08)
-* Additional data about transaction
+* Additional data about transaction	
 * Dynamic templates in loyalty transaction
 
 <a name="0.6.0"></a>